# spring/bin/main/openapi.yml
openapi: 3.0.3
info:
  title: IF Risk Assessment API
  description: 고령자 노동·돌봄 판단 시스템 API
  version: 0.1.0

# ✅ [프론트] 이 서버 URL로 API 호출
# ✅ [백엔드] 로컬 실행 포트/도메인 바뀌면 여기 수정 or 환경별 servers 추가
servers:
  - url: http://localhost:8080
    description: Local Server
paths:
...
  /api/applicants/{applicantId}/assessments:
    post:
      tags:
        - Assessment

      # ✅ [프론트] "신청자 기준으로 판단 생성" 버튼 눌렀을 때 호출
      # ✅ [백엔드] AssessmentController.createAssessment(@PathVariable applicantId, @RequestBody ...)
      summary: 신규 판단 생성
      operationId: createAssessment

      parameters:
        # ✅ [프론트] 이 API는 applicantId를 path로 요구함:
        #    POST /api/applicants/1/assessments
        # ✅ [백엔드] applicantId는 @PathVariable Long applicantId 로 수신
        - name: applicantId
          in: path
          required: true
          description: 신청자 ID
          schema:
            type: integer
            format: int64

      requestBody:
        required: true
        content:
          application/json:
            schema:
              # ✅ [프론트] jobId, healthId를 body로 넣어야 함
              # ✅ [백엔드] adminId는 body로 받지 않음(인증 도입 시 토큰에서 꺼냄)
              $ref: "#/components/schemas/AssessmentCreateRequest"

      responses:
        "201":
          description: 판단 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
components:
  schemas:
...
    AssessmentCreateRequest:
      type: object
      required:
        - jobId
        - healthId
      properties:
        jobId:
          type: integer
          format: int64
          description: 직무 ID
          example: 101

          # ✅ [프론트] 직무 선택 UI에서 jobId를 선택해서 body로 전송
          # ✅ [백엔드] jobId 유효성(존재 여부) 체크는 service에서 필수

        healthId:
          type: integer
          format: int64
          description: 건강 정보 ID
          example: 202

          # ✅ [프론트] 건강정보 선택 UI에서 healthId 선택
          # ✅ [백엔드] healthId 유효성 체크도 service에서 처리
    AssessmentResponse:
      type: object
      required:
        - id
        - status
        - assessedAt
      properties:
        id:
          type: integer
          format: int64
          description: 판단 ID (External ID)
          example: 1001

          # ✅ [프론트] 판단 상세/상태 조회가 생기면 이 id로 라우팅 가능
          # ✅ [백엔드] PK 노출 정책은 장기적으로 UUID 도입 고려 가능

        status:
          type: string
          description: 판단 상태
          enum:
            - PENDING_AI
            - AI_COMPLETED
            - FINALIZED
          example: PENDING_AI

          # ✅ [프론트] 상태에 따라 UI 표시(뱃지/색상/버튼 활성화) 분기
          # ✅ [백엔드] 상태 전이는 service에서만 변경되게(컨트롤러에서 직접 set 금지)

        assessedAt:
          type: string
          format: date-time
          description: 판단 생성 시각 (ISO-8601, UTC)
          example: 2026-02-10T12:00:00Z

          # ✅ [프론트] createdAt처럼 날짜 포맷 통일해서 표시
          # ✅ [백엔드] 생성 시각/판단 시각 의미가 달라지면 필드 분리 필요(예: createdAt vs finalizedAt)